authorization

+++ routes/posts.js
@@ -35,7 +35,8 @@ router.post('',
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
   post.save().then(result => {
     res.status(201).json({
@@ -96,24 +97,32 @@ router.put('/:id',
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagepath
+      imagePath: imagepath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({_id: req.params.id}, post )
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post )
       .then(documents => {
-        res.status(200).json({
-          message: 'Posts Updated successfully!'
-        });
+        console.log(documents);
+        if(documents.nModified > 0){
+          res.status(200).json({
+            message: 'Posts Updated successfully!'
+          });
+        }else{
+          res.status(401).json({
+            message: 'Not Authorized!!'
+          });
+        }
+
       });
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  console.log(req.params.id);
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({
-      message: 'Post deleted'
-    });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if(result.n > 0){
+      res.status(200).json({ message: 'Post deleted'});
+    }else{
+      res.status(401).json({ message: 'Not Authorized!!'});
+    }
   });
 });
